+++ src/views/dashboard/Dashboard.js
@@ -1,10 +1,12 @@
 import React from 'react'
 
-
 const Dashboard = () => {
+
   return (
     <>
       Dashboard works
+      <div>
+      </div>
     </>
   )
 }