+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Changelog
 
+-   Added a login function with Metamask
 -   Added a basic layout consisting of a dashboard, a sidebar and a navbar